Fix 1.34.0 build and other CI issues
