+++ Cargo.toml
@@ -10,10 +10,11 @@ edition = "2018"
 [features]
 default = ["reqwest"]
 futures-01 = ["futures-0-1"]
-futures-03 = ["futures-0-3", "async-trait", "reqwest-0-10"]
+futures-03 = ["futures-0-3", "async-std", "async-trait", "reqwest-0-10"]
 reqwest = ["reqwest-0-9"]
 
 [dependencies]
+async-std = { version = "1.0", optional = true }
 async-trait = { version = "0.1", optional = true }
 base64 = "0.10"
 curl = { version = "0.4.0", optional = true }
@@ -31,5 +32,4 @@ sha2 = "0.7"
 url = "2.1"
 
 [dev-dependencies]
-async-std = "1.0"
 tokio = "0.1"
